backend/requirements.txt
fastapi
uvicorn
python-multipart
pytesseract
pdfplumber
python-docx
python-pptx
Pillow
pymupdf
